feat: add parser toml for config
