+++ vsmtp-config/src/next/mod.rs
@@ -1,5 +1,5 @@
-// #[cfg(test)]
-// mod tests;
+#[cfg(test)]
+mod tests;
 
 mod parser {
     pub mod semver;
@@ -9,10 +9,20 @@ mod parser {
     pub mod tls_protocol_version;
 }
 
-mod builder;
+mod builder {
+    ///
+    pub mod validate;
+
+    ///
+    pub mod wants;
+
+    ///
+    pub mod with;
+}
+
 mod config;
 
-pub use builder::{Builder, WantsVersion};
+pub use builder::{validate, wants::*, with::*};
 pub use config::Config;
 
 impl Config {
@@ -22,5 +32,41 @@ impl Config {
         Builder {
             state: WantsVersion(()),
         }
+    }
+
+    /// Parse a [ServerConfig] with [TOML] format
+    ///
+    /// # Errors
+    ///
+    /// * data is not a valid [TOML]
+    /// * one field is unknown
+    /// * the version requirement are not fulfilled
+    /// * a mandatory field is not provided (no default value)
+    ///
+    /// [TOML]: https://github.com/toml-lang/toml
+    pub fn from_toml(input: &str) -> anyhow::Result<Self> {
+        #[derive(serde::Serialize, serde::Deserialize)]
+        struct VersionRequirement {
+            #[serde(
+                serialize_with = "crate::next::parser::semver::serialize",
+                deserialize_with = "crate::next::parser::semver::deserialize"
+            )]
+            version_requirement: semver::VersionReq,
+        }
+
+        let req = toml::from_str::<VersionRequirement>(input)?;
+        let pkg_version = semver::Version::parse(env!("CARGO_PKG_VERSION"))?;
+
+        if !req.version_requirement.matches(&pkg_version) {
+            anyhow::bail!(
+                "Version requirement not fulfilled: expected '{}' but got '{}'",
+                req.version_requirement,
+                env!("CARGO_PKG_VERSION")
+            );
+        }
+
+        toml::from_str::<Self>(input)
+            .map_err(anyhow::Error::new)
+            .map(Builder::<WantsValidate>::ensure)
     }
 }
