+++ vsmtp-config/src/next/tests/from_toml/mod.rs
@@ -1,0 +1,1 @@
+mod one;
