+++ vsmtp-config/src/next/tests/from_toml/one.rs
@@ -1,0 +1,7 @@
+use crate::next::Config;
+
+#[test]
+fn parse() {
+    let config = Config::from_toml(include_str!("one.toml"));
+    assert!(config.is_ok(), "{:?}", config);
+}
