+++ vsmtp-config/src/next/tests/from_toml/one.toml
@@ -1,0 +1,114 @@
+version_requirement = ">=0.9.5, <1.0.0"
+
+[server]
+domain = "testserver.com"
+client_count_max = 16
+
+[server.system]
+user = "vsmtp"
+group = "vsmtp"
+
+[server.system.thread_pool]
+receiver = 6
+processing = 6
+delivery = 6
+
+[server.interfaces]
+addr = ["0.0.0.0:25"]
+addr_submission = ["127.0.0.1:587"]
+addr_submissions = ["10.0.4.254:465"]
+
+[server.logs]
+filepath = "/var/log/vsmtp/vsmtp.log"
+format = "{d} {l} - "
+
+[server.logs.level]
+default = "warn"
+receiver = "info"
+rule_engine = "warn"
+delivery = "error"
+
+
+[server.queues]
+dirpath = "/var/spool/vsmtp"
+
+[server.queues.working]
+channel_size = 32
+
+[server.queues.delivery]
+channel_size = 32
+deferred_retry_max = 10
+deferred_retry_period = "10s"
+# dead_file_lifetime = "14d"
+
+
+[server.tls]
+security_level = "May"
+preempt_cipherlist = true
+handshake_timeout = "100ms"
+protocol_version = "^TLSv1.2"
+certificate = "../tmp/config/certs/certificate.crt"
+private_key = "../tmp/config/certs/privateKey.key"
+
+[[server.tls.sni]]
+domain = "testserver2.com"
+certificate = "../tmp/config/certs/certificate.crt"
+private_key = "../tmp/config/certs/privateKey.key"
+
+[[server.tls.sni]]
+domain = "testserver3.com"
+certificate = "../tmp/config/certs/certificate.crt"
+private_key = "../tmp/config/certs/privateKey.key"
+
+
+[server.smtp]
+rcpt_count_max = 25
+disable_ehlo = false
+required_extension = ["STARTTLS", "SMTPUTF8", "8BITMIME", "AUTH"]
+
+[server.smtp.error]
+soft_count = 5
+hard_count = 10
+delay = "5000ms"
+
+[server.smtp.timeout_client]
+connect = "50ms"
+helo = "100ms"
+mail_from = "200ms"
+rcpt_to = "400ms"
+data = "800ms"
+
+
+[server.smtp.codes]
+Code214 = "214 my custom help message\r\n"
+Code220 = "220 {domain} ESMTP Service ready\r\n"
+
+
+# [server.smtp.startls]
+# ???
+
+
+[app]
+dirpath = "/var/spool/vsmtp/app"
+
+[app.vsl]
+filepath = "/etc/vsmtp/main.vsl"
+
+[app.logs]
+filepath = "/var/log/vsmtp/app.log"
+level = "trace"
+format = "{d} - {m}{n}"
+
+[app.services.clamscan]
+type = "shell"
+timeout = "15s"
+command = "/etc/vsmtp/service/clamscan.sh"
+args = "{mail}"
+
+[app.services.spamassassin]
+type = "shell"
+timeout = "15s"
+command = "/etc/vsmtp/service/spamassassin.sh"
+args = "{mail}"
+user = "spamassassin"
+group = "spamassassin"
